+++ choco-solver/src/main/java/solver/constraints/binary/PropNotEqualX_Y.java
@@ -27,7 +27,6 @@
 
 package solver.constraints.binary;
 
-import choco.annotations.PropAnn;
 import solver.constraints.Propagator;
 import solver.constraints.PropagatorPriority;
 import solver.exception.ContradictionException;
@@ -58,7 +57,6 @@ import util.tools.ArrayUtils;
  * @version 0.01, june 2010
  * @since 0.01
  */
-@PropAnn(tested = PropAnn.Status.EXPLAINED)
 public class PropNotEqualX_Y extends Propagator<IntVar> {
 
     IntVar x;