+++ CHANGES.md
@@ -78,3 +78,4 @@ The CHANGES.md file starts at this version.
 * Stronger AtMostNValue constraint with automatic detection of disequalities 
 * Fix #114: Enable to specify a time unit in time limits (ms, s, m or h)
 * Fix #195: fix bug while using IntViews over BoolVar
+* Fix #17: propagator annotations (PropAnn) have been removed